requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7 